+++ backend/src/main/java/com/team5/issue_tracker/user/query/UserQueryService.java
@@ -15,12 +15,12 @@ import lombok.extern.slf4j.Slf4j;
 public class UserQueryService {
   private final UserQueryRepository userQueryRepository;
 
-  public UserScrollResponse getScrolledUsers(String cursor, Long limit) {
+  public UserScrollResponse getScrolledUsers(String cursor, Integer limit) {
     log.debug("유저 스크롤 조회 요청");
     List<UserSummaryResponse> usersPlusOne = userQueryRepository.getScrolledUsers(cursor, limit);
 
     Boolean hasNext = usersPlusOne.size() > limit;
-    List<UserSummaryResponse> users = hasNext ? usersPlusOne.subList(0, limit.intValue()) : usersPlusOne;
+    List<UserSummaryResponse> users = hasNext ? usersPlusOne.subList(0, limit) : usersPlusOne;
     String nextCursor = hasNext ? users.getLast().getUsername() : null;
 
     return new UserScrollResponse(hasNext, nextCursor, users);
